backend/requirements.txt
# Core Framework
fastapi
uvicorn[standard]

# Database
motor           # Async MongoDB driver
dnspython       # Required for MongoDB Atlas SRV URIs

# Auth & Security
python-jose[cryptography]
bcrypt
passlib[bcrypt]

# Config & Utilities
python-dotenv
python-multipart
pydantic[email]
email-validator

# HTTP & Networking
httpx
anyio
